requirements.txt
# core
Werkzeug
jinja2
WTForms
flask_wtf

# DB
#flask_migrate#==4.1.0
#flask_sqlalchemy#==3.1.1
#sqlalchemy

# tools
#flask_login
flask-dance
celery
redis
colorama
PyJWT
WTForms-Alchemy

# utils
email_validator
blinker

# env
python-dotenv

# deployment
gunicorn
Flask-Minify
Flask-CDN

# flask_mysqldb
# psycopg2-binary
